backend/app/agents/orchestrator/agent.py
"""
Orchestrator Agent - Web3 Agent Coordination Agent

This module implements an AI-powered orchestrator agent that coordinates
specialized A2A agents for Web3 and cryptocurrency operations. It receives
user requests via AG-UI Protocol and delegates tasks to appropriate
specialized agents.

ARCHITECTURE OVERVIEW:
----------------------
The orchestrator follows a coordination pattern:

1. AG-UI ADK Layer:
   - Uses Google ADK's LlmAgent with Gemini 2.5 Pro model
   - Exposed via AG-UI Protocol through ADKAgent wrapper
   - Uses FastAPI with add_adk_fastapi_endpoint for HTTP interface
   - Supports session management and in-memory services

2. Agent Coordination Layer:
   - Receives user queries and determines which specialized agent to call
   - Currently coordinates with Balance Agent (A2A protocol)
   - Uses send_message_to_a2a_agent tool (provided by frontend middleware)
   - Enforces sequential agent calls (no parallel/concurrent calls)

3. Specialized Agents:
   - Balance Agent: Checks cryptocurrency balances across multiple chains
   - Future agents can be added (transfer, swap, etc.)

WORKFLOW:
---------
1. User sends a query (e.g., "Check my USDC balance on Ethereum")
2. Orchestrator agent receives query via AG-UI Protocol
3. Agent analyzes query and determines required specialized agent
4. Agent calls Balance Agent (or other specialized agent) via A2A protocol
5. Waits for response from specialized agent
6. Formats and presents results to user
7. For multiple queries, processes sequentially (one at a time)

KEY COMPONENTS:
---------------
- LlmAgent: Core Gemini-based agent with detailed orchestration instructions
- ADKAgent: Wraps LlmAgent for AG-UI Protocol compatibility
- create_orchestrator_agent_app(): Factory function to create FastAPI app

ENVIRONMENT VARIABLES:
----------------------
- GOOGLE_API_KEY: Required - Google AI Studio API key for Gemini model access

USAGE:
------
Mounted mode (recommended):
    from app.agents.orchestrator.agent import create_orchestrator_agent_app
    app.mount("/orchestrator", create_orchestrator_agent_app())

NOTES:
------
- Uses Gemini 2.5 Pro model for orchestration logic
- Enforces sequential agent calls (critical constraint)
- Validates wallet addresses (0x format, 42 characters)
- Supports multiple EVM chains (Ethereum, BNB, Polygon, etc.)
- Uses in-memory services (sessions, artifacts, memory) - not persistent
- Frontend A2A middleware provides send_message_to_a2a_agent tool
"""

from __future__ import annotations

from dotenv import load_dotenv

load_dotenv()

from fastapi import FastAPI
from ag_ui_adk import ADKAgent, add_adk_fastapi_endpoint
from google.adk.agents import LlmAgent


orchestrator_agent = LlmAgent(
    name="OrchestratorAgent",
    model="gemini-2.5-pro",
    instruction="""
    You are a Web3 and cryptocurrency orchestrator agent. Your role is to coordinate
    specialized agents to help users with blockchain and cryptocurrency operations.

    AVAILABLE SPECIALIZED AGENTS:

    1. **Balance Agent** (LangGraph) - Checks cryptocurrency balances across multiple chains
       - Supports Ethereum, BNB, Polygon, and other EVM-compatible chains
       - Can check native token balances (ETH, BNB, MATIC, etc.)
       - Can check ERC-20 token balances (USDC, USDT, DAI, etc.)
       - Requires wallet address (0x format) and optional network specification

    CRITICAL CONSTRAINTS:
    - You MUST call agents ONE AT A TIME, never make multiple tool calls simultaneously
    - After making a tool call, WAIT for the result before making another tool call
    - Do NOT make parallel/concurrent tool calls - this is not supported
    - Always validate wallet addresses are in 0x format and 42 characters long

    RECOMMENDED WORKFLOW FOR CRYPTO OPERATIONS:

    1. **Balance Agent** - Check cryptocurrency balances
       - Extract wallet address from user query (format: 0x...)
       - Extract network if specified (ethereum, bnb, polygon, etc.) - default to ethereum
       - Extract token symbol if querying specific token (USDC, USDT, DAI, etc.)
       - Call Balance Agent with appropriate parameters:
         * For native balance: address and network
         * For token balance: address, token symbol, and network
       - Wait for balance response
       - Present results in a clear, user-friendly format

    WORKFLOW EXAMPLES:

    Example 1: Simple balance check
    - User: "Check my balance"
    - Extract: Ask for wallet address if not provided
    - Call Balance Agent: address, network="ethereum" (default)
    - Present: Native ETH balance

    Example 2: Multi-chain balance
    - User: "Get my balance on Polygon"
    - Extract: address (if provided), network="polygon"
    - Call Balance Agent: address, network="polygon"
    - Present: Native MATIC balance

    Example 3: Token balance
    - User: "Check my USDC balance on Ethereum"
    - Extract: address, token="USDC", network="ethereum"
    - Call Balance Agent: address, token="USDC", network="ethereum"
    - Present: USDC token balance

    Example 4: Multiple queries
    - User: "Check my ETH balance and USDT balance on BNB"
    - First call: Balance Agent for ETH on BNB
    - Wait for result
    - Second call: Balance Agent for USDT on BNB
    - Wait for result
    - Present: Combined results

    ADDRESS VALIDATION:
    - Wallet addresses must start with "0x" and be 42 characters long
    - If user provides invalid address, politely ask for correct format
    - If address is missing, ask user to provide it

    NETWORK SUPPORT:
    - Ethereum (default): ethereum, eth
    - BNB Chain: bnb, bsc, binance
    - Polygon: polygon, matic
    - Other EVM chains as supported by Balance Agent

    TOKEN SUPPORT:
    - Common tokens: USDC, USDT, DAI, WBTC, WETH
    - Token symbols are case-insensitive
    - Always use uppercase for token symbols in responses

    RESPONSE STRATEGY:
    - After each agent response, acknowledge what you received
    - Format balance results clearly with:
      * Network name
      * Token symbol (if applicable)
      * Balance amount with appropriate decimals
      * Wallet address (truncated for display: 0x...last4)
    - For multiple queries, organize results by network or token type
    - If there's an error, explain it clearly and suggest alternatives

    IMPORTANT: Once you have received a response from an agent, do NOT call that same
    agent again for the same information. Use the information you already have.

    ERROR HANDLING:
    - If balance check fails, explain the error clearly
    - Suggest checking: address format, network availability, token contract address
    - For network errors, suggest trying a different network or checking connectivity
    """,
)


def create_orchestrator_agent_app() -> FastAPI:
    """Create and configure the AG-UI ADK application for the orchestrator agent.

    Returns:
        FastAPI application instance configured for the orchestrator agent
    """
    # Expose the agent via AG-UI Protocol
    adk_orchestrator_agent = ADKAgent(
        adk_agent=orchestrator_agent,
        app_name="orchestrator_app",
        user_id="demo_user",
        session_timeout_seconds=3600,
        use_in_memory_services=True,
    )

    app = FastAPI(title="Web3 Orchestrator Agent (ADK)")
    add_adk_fastapi_endpoint(app, adk_orchestrator_agent, path="/")
    return app
