backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "movement-backend"
version = "0.1.0"
description = "Backend API for Movement Hackathon"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.16",
    "ag-ui-adk>=0.3.4",
    "fastapi>=0.104.1",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.2.1",
    "python-multipart==0.0.12",
    "langchain-core",
    # Blockchain dependencies
    "web3>=6.15.0",
    "requests>=2.32.5",
    # Google ADK for liquidity agent
    "google-adk>=1.17.0",
    # Token research dependencies
    "coingecko>=0.13.0",
    "duckduckgo-search>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "black==24.8.0",
    "ruff==0.6.9",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "httpx>=0.28.1",
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

